Move SwaggerUiBuildItem to an spi project so it can be accessed by other extensions
