+++ extensions/swagger-ui/pom.xml
@@ -17,6 +17,7 @@
     <modules>
         <module>deployment</module>
         <module>runtime</module>
+        <module>spi</module>
     </modules>
 
 </project>
